Move Google Sheets related files into one folder
